replace Time.time with Time.timeAsDouble - fix problems with floating-point numbers precision when game uptime is high (eg. 1 month)

+++ Assets/Scripts/Utilities/IState.cs
@@ -15,8 +15,8 @@ namespace SanAndreasUnity.Utilities
 
 		object ParameterForEnteringState { set; }
 
-		float LastTimeWhenActivated { get; set; }
-		float LastTimeWhenDeactivated { get; set; }
+		double LastTimeWhenActivated { get; set; }
+		double LastTimeWhenDeactivated { get; set; }
 	}
 
 }
